修复zk超时引起monitor一些bug

+++ dubbo-test/dubbo-test-benchmark/src/main/java/com/alibaba/dubbo/rpc/benchmark/AbstractClientRunnable.java
@@ -62,7 +62,7 @@ public abstract class AbstractClientRunnable implements ClientRunnable {
         serviceFactory.setClientNums(clientNums);
         serviceFactory.setTargetPort(targetPort);
         serviceFactory.setConnectTimeout(rpcTimeout);
-        maxRange = (Integer.parseInt(String.valueOf((endTime - startTime))) / 1000000) + 1;
+        maxRange = ((int)(Long.parseLong(String.valueOf((endTime - startTime))) / 1000000L)) + 1;
         errorTPS = new long[maxRange];
         errorResponseTimes = new long[maxRange];
         tps = new long[maxRange];
